Octo test/requirements.txt
praw
requests
matplotlib
python-dotenv
